Provide NewInstanceSetSourceForApp method
Allow use of a Eureka application as a source of instances just like a
VIP address, optionally subjected to the same filtering and shuffling.
