+++ example_vipupdate_test.go
@@ -76,58 +76,3 @@ func ExampleEurekaConnection_ScheduleSecureVIPAddressUpdates_context() {
 	}
 	fmt.Printf("Done monitoring secure VIP address %q.\n", svipAddress)
 }
-
-func ExampleInstanceSetSource_Latest_outcomes() {
-	e := makeConnection()
-	vipAddress := "my_vip"
-	source, err := e.NewInstanceSetSourceForVIPAddress(vipAddress, false, fargo.ThatAreUp)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	defer source.Stop()
-	time.Sleep(30 * time.Second)
-	if instances := source.Latest(); instances != nil {
-		fmt.Printf("VIP address %q has %d instances available.\n", vipAddress, len(instances))
-	}
-	time.Sleep(time.Minute)
-	switch instances := source.Latest(); {
-	case instances == nil:
-		fmt.Printf("Unsure whether any instances for VIP address %q are available.\n", vipAddress)
-	case len(instances) == 0:
-		fmt.Printf("No instances for VIP address %q are available.\n", vipAddress)
-	default:
-		fmt.Printf("VIP address %q has %d instances available.\n", vipAddress, len(instances))
-	}
-}
-
-func ExampleInstanceSetSource_Latest_compare() {
-	e := makeConnection()
-	svipAddress := "my_vip"
-	source, err := e.NewInstanceSetSourceForSecureVIPAddress(svipAddress, true, fargo.WithStatus(fargo.DOWN), fargo.WithStatus(fargo.OUTOFSERVICE))
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	defer source.Stop()
-	var troubled []*fargo.Instance
-	for remaining := 10; ; {
-		if instances := source.Latest(); instances != nil {
-			fmt.Printf("VIP address %q has %d troubled instances.", svipAddress, len(instances))
-			switch diff := len(instances) - len(troubled); {
-			case diff > 0:
-				fmt.Printf(" (%d more than last time)\n", diff)
-			case diff < 0:
-				fmt.Printf(" (%d fewer than last time)\n", diff)
-			default:
-				fmt.Println()
-			}
-			troubled = instances
-		}
-		remaining--
-		if remaining == 0 {
-			break
-		}
-		time.Sleep(30 * time.Second)
-	}
-}
